Set up the world generator interface for use with the API

+++ api/src/main/java/com/seibel/lod/api/enums/EDhApiDetailLevel.java
@@ -40,13 +40,13 @@ public enum EDhApiDetailLevel
 	REGION(9, 512);
 	
 	
-	public final int detailLevel;
-	public final int widthInBlocks;
+	public final byte detailLevel;
+	public final byte widthInBlocks;
 	
 	EDhApiDetailLevel(int detailLevel, int widthInBlocks)
 	{
-		this.detailLevel = detailLevel;
-		this.widthInBlocks = widthInBlocks;
+		this.detailLevel = (byte) detailLevel;
+		this.widthInBlocks = (byte) widthInBlocks;
 	}
 	
 }
